OnlineShop/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
sqlparse==0.4.2
